docker/requirements.txt
ImageHash==4.3.1
joblib==1.1.0
#mediapipe==0.9.1.0
numpy==1.22.4
opencv_python==4.6.0.66
paddleocr==2.6.0.1
pandas==1.4.2
Pillow==9.4.0
#PyQt5==5.15.9
pytube==12.1.2
#QPT==1.0b4.dev6
setuptools==67.6.1
spacy==3.3.1
boto3==1.26.98
python-pptx==0.6.21
scikit-learn==1.1.1
streamlit